pass user-defined bias argument to bootstrap estimations
in Bootstrap::run_em, EMAlgorithm::run is called with the fourth
argument set to false, that is without bias estimation even if the
user set the --bias option on command line. This causes in some cases
major discrepancies between the intial EM estimate and bootstrap
estimates, which renders the bootstrap estimates inadequate to
estimate the uncertainty of the abundance ML estimates (see #151 for
an example).